Bare-bones GUI trainer
