+++ nam/train/_version.py
@@ -1,0 +1,24 @@
+# File: _version.py
+# Created Date: Tuesday December 20th 2022
+# Author: Steven Atkinson (steven@atkinson.mn)
+
+"""
+Version utility
+"""
+
+class Version:
+    def __init__(self, major: int, minor: int, patch: int):
+        self.major = major
+        self.minor = minor
+        self.patch = patch
+
+    def __lt__(self, other) -> bool:
+        if self.major != other.major:
+            return self.major < other.major
+        if self.minor != other.minor:
+            return self.minor < other.minor
+        if self.patch != other.patch:
+            return self.patch < other.patch
+
+    def __str__(self) -> str:
+        return f"{self.major}.{self.minor}.{self.patch}"
